% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{calc_rb}
\alias{calc_rb}
\title{Calculate Relative Bias}
\usage{
calc_rb(estimates, true_val)
}
\arguments{
\item{estimates}{A numeric vector of parameter estimates.}

\item{true_val}{The true parameter value.}
}
\value{
The relative bias in percentage.
}
\description{
Calculate Relative Bias
}
